Conflicts:
	mysql-test/r/dd_schema_definition_debug_cs.result
	mysql-test/suite/innodb/include/innodb-system-table-view.inc
	mysql-test/suite/innodb/t/disabled.def
	mysql-test/suite/innodb/t/innodb-index-online-fk.test
	sql/item_create.cc
	storage/innobase/dict/dict0crea.cc
	storage/innobase/dict/dict0dd.cc
	storage/innobase/handler/ha_innodb.cc
	storage/innobase/handler/handler0alter.cc
	storage/innobase/include/dict0mem.h
	storage/innobase/pars/pars0pars.cc

+++ rapid/plugin/group_replication/include/certifier.h
@@ -1,4 +1,4 @@
-/* Copyright (c) 2014, 2016, Oracle and/or its affiliates. All rights reserved.
+/* Copyright (c) 2014, 2017, Oracle and/or its affiliates. All rights reserved.
 
    This program is free software; you can redistribute it and/or modify
    it under the terms of the GNU General Public License as published by
@@ -16,21 +16,21 @@
 #ifndef CERTIFIER_INCLUDE
 #define CERTIFIER_INCLUDE
 
-#include <map>
-#include <unordered_map>
-#include <string>
-#include <list>
-#include <vector>
-
-#include "certifier_stats_interface.h"
-#include "member_info.h"
-#include "gcs_plugin_messages.h"
-#include "plugin_utils.h"
-#include "pipeline_interfaces.h"
-
 #include <mysql/gcs/gcs_communication_interface.h>
 #include <mysql/gcs/gcs_control_interface.h>
 #include <mysql/group_replication_priv.h>
+#include <list>
+#include <map>
+#include <string>
+#include <unordered_map>
+#include <vector>
+
+#include "certifier_stats_interface.h"
+#include "gcs_plugin_messages.h"
+#include "member_info.h"
+#include "my_dbug.h"
+#include "pipeline_interfaces.h"
+#include "plugin_utils.h"
 
 
 /**
@@ -180,7 +180,7 @@ class Certifier_interface : public Certifier_stats
 public:
   virtual ~Certifier_interface() {}
   virtual void handle_view_change()= 0;
-  virtual int handle_certifier_data(const uchar *data, uint len,
+  virtual int handle_certifier_data(const uchar *data, ulong len,
                                     const Gcs_member_identifier& gcs_member_id)= 0;
 
   virtual void get_certification_info(std::map<std::string, std::string> *cert_info)= 0;
@@ -234,7 +234,7 @@ public:
       @retval 0      OK
       @retval !=0    Error on queue
   */
-  virtual int handle_certifier_data(const uchar *data, uint len,
+  virtual int handle_certifier_data(const uchar *data, ulong len,
                                     const Gcs_member_identifier& gcs_member_id);
 
   /**
@@ -329,7 +329,7 @@ public:
   /**
     Get method to retrieve the size of the members.
   */
-  uint get_members_size();
+  size_t get_members_size();
 
   /**
     Generate group GNO for a view change log event.
@@ -390,7 +390,7 @@ public:
     @retval 0    if there is no GTID / the string is empty
     @retval !=0  the size of the string
   */
-  int get_local_certified_gtid(std::string& local_gtid_certified_string);
+  size_t get_local_certified_gtid(std::string& local_gtid_certified_string);
 
   /**
     Enables conflict detection.
@@ -454,7 +454,7 @@ private:
 
     @retval Gtid_set::Interval which is the interval os GTIDs attributed
   */
-  Gtid_set::Interval reserve_gtid_block(long block_size);
+  Gtid_set::Interval reserve_gtid_block(longlong block_size);
 
   /**
     This function updates parallel applier indexes.
